services/cipher_service/requirements.txt
Flask
gunicorn
python-dotenv
flask-cors
cryptography